ref: move geography based helpers to geoHelpers
