+++ src/server/models/listItem/index.ts
@@ -1,1 +1,2 @@
+export * from "./geoHelpers";
 export * from "./listItem";
